+++ Electron/src/pages/dashboard/tracers/tracersDisplay.tsx
@@ -1,54 +1,17 @@
 import React from 'react';
 import DashboardContainer from '../dashboardContainer';
-import Chart from "chart.js/auto";
-import { Bar } from "react-chartjs-2";
-import {CategoryScale} from 'chart.js';
-Chart.register(CategoryScale);
+import BarChart from './barChart';
+
 
 export default function TracersDisplay() {
   return (
-    <div className='w-full'>
+    <div className='overflow-hidden w-full'>
       <DashboardContainer title='Latency'/>
-      <BarChart />
+      <BarChart communications={comms} timeAtoA={fullTime} timeBtoA={returnTime}/>
     </div>
-
   );
 }
 
-const BarChart = () => {
-  const labels = ["Service A", "Service B", "Service A", "Service B", "Service C", "Service B"];
-  const data = {
-    labels: labels,
-    datasets: [
-      {
-        label: "My First dataset",
-        backgroundColor: "rgb(255, 99, 132)",
-        borderColor: "rgb(255, 255, 255)",
-        data: [5, 10, 5, 2, 20, 10],
-      },
-    ],
-  };
-  const options = {
-    indexAxis: 'y' as const,
-    elements: {
-      bar: {
-        borderWidth: 2,
-      },
-    },
-    responsive: true,
-    plugins: {
-      legend: {
-        position: 'right' as const,
-      },
-      title: {
-        display: true,
-        text: 'Chart.js Horizontal Bar Chart',
-      },
-    },
-  };
-  return (
-    <div className='flex'>
-      <Bar options={options} data={data} className='object-fit shrink'/>
-    </div>
-  );
-};
+const comms = ["Service A", "Service B", "Service A", "Service B", "Service C", "Service B", "Service A", "Service B", "Service A", "Service B", "Service C", "Service B", "Service A", "Service B", "Service A", "Service B", "Service C", "Service B"];
+const fullTime = [70, 180, 90, 250, 310, 100, 70, 180, 90, 250, 310, 100, 70, 1800, 90, 250, 310, 100];
+const returnTime = [40, 70, 50, 20, 120, 50, 40, 70, 50, 20, 120, 50, 40, 70, 50, 20, 120, 50];
